Documentation, doctests, and NewSession trait
Add lots of documentation and examples, as well as refactoring the way
to instantiate a channel into a NewSession: Session extension trait
which lets you call a static method on a session type to create a
channel. This also adds unsatisfiable instances for Transmit and Receive
for all Unavailable things, which means that users can't instantiate
that class themselves.