Add constants and types for small numbers
This adds a unary::types and unary::constants pair of modules which
provide from _0 through _127 as types and constants, respectively, for
their peano types/values. This also changes `choose` to take a singleton
peano unary number as a value, not as a turbofish type -- it's always
mandatory, so it's syntactically cleaner this way.