Eliminate runtime failure at the offer! level
Previously, offer! could throw an error indicating that an out-of-bounds
discriminant was received. Instead of permitting this possibility, we
prevent out-of-bounds discriminants from being sent, and force
receivers, if they are deserializing rather than accepting ownership of
an already-created discriminant, to validate that it is in range.

This introduces the new type Choice<N>, which is guaranteed to contain a
`u8` strictly less than the type-level number `N`. This is what is sent
when choices are made now.

Also included in this commit are documentation improvements and a module
refactor for serde prompted by the inclusion of `Choice`.