Fix bug preventing nested loops from typechecking
Due to an incorrect definition of `Select` that was not inductive,
environments could not be arbitrarily large.

Separately, the definition of `Actionable` for `Recur` was incorrect,
such that every iteration caused the environment to be more deeply
nested.

Fixing these both required adding another `Remainder` type associated
with `Select` that computes the remainder of the environment after we
**pop off** post-`Recur`. Without that popping of the environment, nested loops won't typecheck, no matter how hard you try ;)