Track Done/Continue state in the environment
This alters the behavior of Break and Done so that if the broken-to or
current level of the environment (respectively) is in `Done` rather than
`Continue` state, the `Action` is `Done`. This change paves the way for
the new `Do` construct, which will rely on this behavior to build the
"semicolon" operator from the Context-Free Session Types paper.