Alter split/seq API to use "drop smuggling"
Instead of mandating that the closures passed to seq (and now split,
since this is now a function taking a closure, rather than direct-style)
always return the exact identical channel they were passed, we mandate
that they run the channel to completion *and then drop it*. which
secretly smuggles out the tx/rx contained within to the outer context,
where it can be re-used. This means things are more succinct to write,
and it doesn't change the soundness of anything.

This commit implements a number of changes:
- a new `SessionIncomplete` error
- many documentation changes and improvements
- elimination of `unsplit_with`
- a new `over` method in `NewSession`
- `Tx` and `Rx` must now be `Send` and `'static`! They can no longer be
  references! This turned out to be an unused feature, because you
  almost always want threading capability anyhow, and tracking lifetimes
  became inexpressible due to Rust's limitations on HRTBs