Add microbenchmarks measuring Dialectic's mpsc vs raw mpsc
These microbenchmarks currently indicate that Dialectic is 2x slower
than using raw tokio::sync::mpsc::UnboundedChannel<Box<dyn Any + Send>>.
However, they show a throughput of ~10 Melem/s for .recv() and ~8.5
Melem/s for .send(), which is fast enough that this slowdown should not
be a bottleneck in meaningful workloads. It's unclear at this time what
the slowdown is due to, but likely it is because Transmit/Receive must
use boxed futures, so there is an extra pointer indirection in a tight
loop.