Backend refactor number 2: removing {send/recv}_choice, adding macros
This alteration removes the necessity of asking backend implementors to specially treat `Choice`,
which can lead to code duplication. Instead, new functionality in `call-by` version 0.2.0 is used to
send `Choice` values by val/ref/mut as needed by the underlying transport. However, this comes at
the cost of it being more difficult to write out the bounds on `Tx` and `Rx` in generic situations,
so two new procedural macros are introduced: `Receiver` and `Transmitter`, which tack the correct
bounds onto the `where`-clause of the item they are attached to.