Break up Transmitter/Receiver to split out Choice-handling
This enables us to bound every Chan as having a Transmitter and Receiver, because now Unavailable
can implement those traits. And that in turn allows us to have a unified error type which
encompasses sending and receiving errors on channels. And that in turn allows us to implement
auto-flushing, only when it is needed by the protocol. All this, is implemented in this commit.